app/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pyjwt
bcrypt